refactor: split vue modal shell into shared parts
